fix: websocket reuse db conn pool
